chore: Fake 객체를 테스트 모듈로 이동/사용하지 않는 파일 삭제
