+++ app/src/test/java/woowacourse/shopping/FakeCartRepository.kt
@@ -1,5 +1,6 @@
-package woowacourse.shopping.data.cart
+package woowacourse.shopping
 
+import woowacourse.shopping.data.cart.CartRepository
 import woowacourse.shopping.data.cart.entity.CartItem
 import woowacourse.shopping.model.Quantity
 import java.lang.IllegalArgumentException
@@ -22,7 +23,7 @@ class FakeCartRepository(savedCartItems: List<CartItem> = emptyList()) : CartRep
 
     override fun decreaseQuantity(productId: Long) {
         val oldCartItem = cart.find { it.productId == productId }
-        oldCartItem ?: throw IllegalArgumentException(CANNOT_DELETE_MESSAGE)
+        oldCartItem ?: throw IllegalArgumentException()
 
         cart.remove(oldCartItem)
         if (oldCartItem.quantity.count == 1) {
@@ -47,12 +48,12 @@ class FakeCartRepository(savedCartItems: List<CartItem> = emptyList()) : CartRep
 
     override fun deleteCartItem(productId: Long) {
         val deleteCartItem =
-            cart.find { it.productId == productId } ?: throw IllegalArgumentException(CANNOT_DELETE_MESSAGE)
+            cart.find { it.productId == productId } ?: throw IllegalArgumentException()
         cart.remove(deleteCartItem)
     }
 
     override fun find(productId: Long): CartItem {
-        return cart.find { it.productId == productId } ?: throw IllegalArgumentException(CANNOT_FIND_MESSAGE)
+        return cart.find { it.productId == productId } ?: throw IllegalArgumentException()
     }
 
     override fun findRange(
@@ -65,9 +66,4 @@ class FakeCartRepository(savedCartItems: List<CartItem> = emptyList()) : CartRep
     }
 
     override fun totalCartItemCount(): Int = cart.size
-
-    companion object {
-        private const val CANNOT_DELETE_MESSAGE = "삭제할 수 없습니다."
-        private const val CANNOT_FIND_MESSAGE = "해당하는 장바구니 상품이 존재하지 않습니다."
-    }
 }
